Add files via upload
Photoacoustic (PA) signals encode information about the optical absorption and spatial distribution of absorbing chromophores as well as the light distribution in the medium. The wavelength dependence of the latter affects the accuracy in chromophore quantification, including estimations of oxygen saturation (sO2) with depth. We propose the use of the ratio of the PA radiofrequency (RF) spectral slopes (SS) at different optical wavelengths to generate frequency filters which can be used to match the fluence profiles across separate images generated with different optical wavelengths. 
Proof-of-principle experiments were carried on a plastic tube with blood of a known oxygenation inserted into a porcine tissue. The algorithm was tested in-vivo in the hind leg of six CD1 mice, each under three different breathing conditions (100% O2, room air and 100% CO2). Imaging was done using the VevoLAZR system at the wavelengths 720 and 870 nm. The SS was calculated from the linear fit of the ratio of the photoacosutic RF power spectra at the two wavelengths. An ultrasound frequency filter was designed and applied to each segmented PA signal in the frequency domain and inversely transformed into the time domain to correct for the differences in the fluence profiles at both wavelengths. Linear spectral unmixing was used to estimate sO2 before and after applying the ultrasound frequency filter. 
The estimated blood sO2 in the plastic tube for the porcine tissue experiment improved by 10.3% after applying the frequency filter when compared to the sO2 measured by a blood gas analyzer. For the in-vivo mouse experiments, the applied sO2 correction was 2.67, 1.33 and -3.33 % for every mm of muscle tissue for mice breathing 100% O2, room air and 100% CO2, respectively. The approach presented here provides a new approach for fluence matching that can potentially improve the accuracy of sO2 estimates by removing the fluence depth dependence at different optical wavelengths.